Rename Apache.Arrow.Variant to Apache.Arrow.Scalars
Rename assembly, project, namespace, and test project:
- src/Apache.Arrow.Variant -> src/Apache.Arrow.Scalars
- test/Apache.Arrow.Variant.Tests -> test/Apache.Arrow.Scalars.Tests
- Update all namespace references across the solution
