+++ Apache.Arrow.Tests.slnf
@@ -7,7 +7,7 @@
       "test\\Apache.Arrow.Flight.Tests\\Apache.Arrow.Flight.Tests.csproj",
       "test\\Apache.Arrow.Flight.TestWeb\\Apache.Arrow.Flight.TestWeb.csproj",
       "test\\Apache.Arrow.Tests\\Apache.Arrow.Tests.csproj",
-      "test\\Apache.Arrow.Variant.Tests\\Apache.Arrow.Variant.Tests.csproj",
+      "test\\Apache.Arrow.Scalars.Tests\\Apache.Arrow.Scalars.Tests.csproj",
       "test\\Apache.Arrow.Operations.Tests\\Apache.Arrow.Operations.Tests.csproj"
     ]
   }
