+++ Apache.Arrow.sln
@@ -1,4 +1,4 @@
-﻿
+
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29926.136
@@ -29,11 +29,11 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Apache.Arrow.Flight.Integra
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Apache.Arrow.IntegrationTest", "test\Apache.Arrow.IntegrationTest\Apache.Arrow.IntegrationTest.csproj", "{E8264B7F-B680-4A55-939B-85DB628164BB}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Apache.Arrow.Variant", "src\Apache.Arrow.Variant\Apache.Arrow.Variant.csproj", "{EE364C9A-CB88-4DCE-9209-6ACBB8E6934F}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Apache.Arrow.Scalars", "src\Apache.Arrow.Scalars\Apache.Arrow.Scalars.csproj", "{EE364C9A-CB88-4DCE-9209-6ACBB8E6934F}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Apache.Arrow.Variant.Tests", "test\Apache.Arrow.Variant.Tests\Apache.Arrow.Variant.Tests.csproj", "{0C495233-010C-45F8-BAB2-D9CD0B7B9861}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Apache.Arrow.Scalars.Tests", "test\Apache.Arrow.Scalars.Tests\Apache.Arrow.Scalars.Tests.csproj", "{0C495233-010C-45F8-BAB2-D9CD0B7B9861}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Apache.Arrow.Variant.Benchmarks", "test\Apache.Arrow.Variant.Benchmarks\Apache.Arrow.Variant.Benchmarks.csproj", "{F3A8B7C6-D5E4-4321-9ABC-DEF012345678}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Apache.Arrow.Scalars.Benchmarks", "test\Apache.Arrow.Scalars.Benchmarks\Apache.Arrow.Scalars.Benchmarks.csproj", "{F3A8B7C6-D5E4-4321-9ABC-DEF012345678}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Apache.Arrow.Operations", "src\Apache.Arrow.Operations\Apache.Arrow.Operations.csproj", "{B4799C3C-179F-470F-A8A4-8A75D172E3E1}"
 EndProject
