+++ src/Apache.Arrow.Operations/Json/VariantJsonReader.cs
@@ -16,7 +16,7 @@
 using System;
 using System.Text;
 using System.Text.Json;
-using Apache.Arrow.Variant;
+using Apache.Arrow.Scalars;
 
 namespace Apache.Arrow.Operations.Json
 {
