+++ src/Apache.Arrow.Operations/Json/VariantJsonWriter.cs
@@ -21,7 +21,7 @@ using System.IO;
 #endif
 using System.Text;
 using System.Text.Json;
-using Apache.Arrow.Variant;
+using Apache.Arrow.Scalars;
 
 namespace Apache.Arrow.Operations.Json
 {
