+++ src/Apache.Arrow.Scalars/Apache.Arrow.Scalars.csproj
