+++ src/Apache.Arrow.Scalars/VariantValueWriter.cs
@@ -21,7 +21,7 @@ using System.Data.SqlTypes;
 using System.IO;
 using System.Text;
 
-namespace Apache.Arrow.Variant
+namespace Apache.Arrow.Scalars
 {
     /// <summary>
     /// Streams variant value bytes directly from primitive calls, without
