+++ test/Apache.Arrow.Operations.Tests/Json/VariantJsonTests.cs
@@ -18,7 +18,7 @@ using System.Collections.Generic;
 using System.Data.SqlTypes;
 using System.Text.Json;
 using Apache.Arrow.Operations.Json;
-using Apache.Arrow.Variant;
+using Apache.Arrow.Scalars;
 using Xunit;
 
 namespace Apache.Arrow.Operations.Tests
