+++ test/Apache.Arrow.Variant.Benchmarks/Apache.Arrow.Variant.Benchmarks.csproj
@@ -10,7 +10,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Apache.Arrow.Variant\Apache.Arrow.Variant.csproj" />
+    <ProjectReference Include="..\..\src\Apache.Arrow.Scalars\Apache.Arrow.Scalars.csproj" />
     <ProjectReference Include="..\..\src\Apache.Arrow.Operations\Apache.Arrow.Operations.csproj" />
   </ItemGroup>
